# [wip,dnm] storage: introduced dedicated raft storage