Implements #16361.

This is a breaking change. To see why consider that prior to this we
stored all consensus data in addition to all system metadata and user
level keys in the same, single RocksDB instance. Here we introduce a
separate, dedicated instance for raft data (log entries and
`HardState`). Cockroach nodes simply restarting with these changes, unless
migrated properly, will fail to find the most recent raft long entries
and `HardState` data in the new RocksDB instance.

Also consider a cluster running mixed versions (nodes with dedicated
raft storage and nodes without), what would the communication between
nodes here like in light of proposer evaluated
KV? Current we propagate a `storagebase.WriteBatch` through raft
containing a serialized representation of a RocksDB write batch, this
models the changes to be made to the single underlying RocksDB instance.
For log truncation requests where we delete log entries and/or admin
splits where we write initial `HardState` for newly formed replicas, we
need to similarly propagate a write batch (through raft) addressing the
new RocksDB instance (if the recipient node is one with these changes)
or the original RocksDB instance (if the recipient node is one without
these changes). What if an older version node is the raft leader and is
therefore the one upstream of raft, propagating `storagebase.WriteBatch`es
with raft data changes but addressed to the original RocksDB instance?
What would rollbacks look like?

To this end we introduce three modes of operation,
`{Disabled,Transitioning,Enabled}RaftStorage`. We've made it so that it is
safe to transition between `DisabledRaftStorage` to
`TransitioningRaftStorage`, from `TransitioningRaftStorage` to
`EnabledRaftStorage` and the reverse for rollbacks. Transition from one
mode to the next will take place when all the nodes in the cluster are
on the same previous mode.  The operation mode is set by an env var
`COCKROACH_DEDICATED_RAFT_STORAGE={DISABLED,TRANSITIONING,ENABLED}`

- `DisabledRaftStorage` mode preserves the previous behavior in that we use
  a single RocksDB instance for both raft and user-level KV data.
- `EnabledRaftStorage` mode enables the use of the dedicated RocksDB
  instance for raft data. Raft log entries and the HardState are stored on
  this instance alone.
- `TransitioningRaftStorage` mode uses both RocksDB instances for raft data
  interoperably, the raft specific and the regular instance. We use
  this mode to facilitate rolling upgrades.

Most of this commit is careful plumbing of an extra
`engine.{Engine,Batch,Reader,Writer,ReadWriter}` for whenever we need to
interact with the new RocksDB instance. In `DisabledRaftStorage` both
these instances refer to the same underlying engine (thus preserving the
previous behaviour). The following pattern is oft repeated:

```go
  batch := ...
  batchRaft := batch
  if TransitioningRaftStorage || EnabledRaftStorage {
    batchRaft = ...
  }
```

Here are some initial performance numbers:

```sh
  ~ benchstat perf-disabled perf-enabled
    name                             old time/op  new time/op  delta
    ReplicaRaftStorage/vs=1024-4      320µs ± 6%   385µs ±18%  +20.29%  (p=0.000 n=10+10)
    ReplicaRaftStorage/vs=4096-4      613µs ±14%   580µs ± 2%     ~     (p=0.278 n=10+9)
    ReplicaRaftStorage/vs=16384-4    2.59ms ± 3%  2.05ms ± 4%  -20.87%  (p=0.000 n=10+9)
    ReplicaRaftStorage/vs=65536-4    4.11ms ± 7%  3.29ms ± 3%  -19.97%  (p=0.000 n=10+10)
    ReplicaRaftStorage/vs=262144-4   13.4ms ± 8%  10.7ms ± 3%  -20.39%  (p=0.000 n=10+10)
    ReplicaRaftStorage/vs=1048576-4  56.8ms ± 3%  36.4ms ± 2%  -35.91%  (p=0.000 n=10+10)
```