@irfansharif 2017-06-30
note: The point above having older versions upstream of raft is as yet unresolved, I have a solution but it's not the cleanest one. Before addressing that I'd like to first know how I can test these cross-version migrations.
re: Tests, insofar my "testing" has been reading from the new engine _and_ the old and verify they're exactly identical. Will add some unit tests but they'll be limited in scope, between the existing raft log + snapshot tests there's little ground left to cover.