@bdarnell 2017-07-02
Yes, the acceptance tests (specifically the ones in `reference_test.go`) are currently our best tests of upgrade-related issues. They use a docker image that contains an old version of the binary (this should probably be refactored so we can use multiple older versions). 

I wonder if it would be better to encapsulate most of this at the Engine level: instead of two Engines, we could have a new implementation of the Engine interface that wraps two other Engines and switches between them based on the keys it is given. 

I had been envisioning the transition as two states instead of three: Whenever a new binary starts up, it is in the "transitioning" state, and it is still possible to roll back to the old version. When the admin changes to the "enabled" state, it is no longer possible to roll back to an older version, although it should be possible to roll back to the "transitioning" state of the new binary. (The old binary is effectively the "disabled" state). The three-state version is nice if we can do it (the fewer restrictions we place on rollbacks, the better), but I worry that it will be tricky to implement and under-tested (once you've transitioned to the "disabled" state, when does it become safe to do the downgrade?)

Store.Start is a good place to run a migration if a restart is required to make the change. We'd prefer not to do that (although it's not out of the question). Instead of a migration that copies data from one engine to the other, I wonder if it would be better to just let the truncation process happen naturally (reading from both engines in the meantime). 

---

Reviewed 29 of 29 files at r1, 1 of 1 files at r2.
Review status: all files reviewed at latest revision, 8 unresolved discussions, some commit checks failed.

---

*[pkg/server/config.go, line 406 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3o3t5KOdmS8Lart2t:-Ko3o3t5KOdmS8Lart2u:bgn0mzr) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/server/config.go#L406)):*
> ```Go
> 
> // CreateEngines creates Engines based on the specs in cfg.Stores.
> func (cfg *Config) CreateEngines(ctx context.Context) (Engines, Engines, error) {
> ```

I'd generally prefer to return an array of (engine, raftEngine) pairs instead of two parallel slices, unless it's too painful to make that refactoring. 

The comment on this method needs to be updated with the new return values.

---

*[pkg/server/config.go, line 411 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3ouAQ9MW2PnHoDoCi:-Ko3ouAQ9MW2PnHoDoCj:b-lsnkz3) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/server/config.go#L411)):*
> ```Go
> 
> 	var raftEngines Engines
> 	if storage.TransitioningRaftStorage || storage.EnabledRaftStorage {
> ```

This repeated expression should perhaps be encapsulated in some global function. 

---

*[pkg/server/server.go, line 683 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3p4TIswDW6NZ56f5H:-Ko3p4TIswDW6NZ56f5I:b3ga5g1) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/server/server.go#L683)):*
> ```Go
> 	}
> 	s.stopper.AddCloser(&s.engines)
> 	if storage.TransitioningRaftStorage || storage.EnabledRaftStorage {
> ```

Refer to the configuration variables as little as possible. Here, for example, you should just close `s.raftEngines` unconditionally. 

As a general rule I'd suggest creating the raft engine (and maybe batches) unconditionally, and the current transition state would only be used to A) decide which batch to write to and B) to skip operations that would impact the overall performance (like committing the write batch). 

---

*[pkg/storage/client_split_test.go, line 1217 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3tZAqByvXpV8TNMUi:-Ko3tZAqByvXpV8TNMUj:b-ftkdiw) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/storage/client_split_test.go#L1217)):*
> ```Go
> func TestSplitSnapshotRace_SnapshotWins(t *testing.T) {
> 	defer leaktest.AfterTest(t)()
> 	t.Skip()
> ```

FYI this test has been flaky on master recently, so it may not be your fault. 

---

*[pkg/storage/client_test.go, line 1003 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3tfqVrZrh1dU3WHQp:-Ko3tfqVrZrh1dU3WHQq:bm1txoj) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/storage/client_test.go#L1003)):*
> ```Go
> 	ctx := context.TODO()
> 	startKey := m.findStartKeyLocked(rangeID)
> 	log.Infof(context.TODO(), "skey: %v", startKey)
> ```

Remember to remove this before merging. 

---

*[pkg/storage/replica_raftstorage.go, line 166 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3uVjhWwyTMHZWkuuC:-Ko3uVjhWwyTMHZWkuuD:b-u5ae27) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/storage/replica_raftstorage.go#L166)):*
> ```Go
> 
> 		// Was the missing index after the last index?
> 		// TODO(irfansharif): Explore writing last index to raft engine.
> ```

Yes, it sounds like a good idea to move the last index to the raft engine. 

---

*[pkg/storage/replica_raftstorage.go, line 180 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3u_CIf_2xfMQiI60G:-Ko3u_CIf_2xfMQiI60H:bhuj78q) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/storage/replica_raftstorage.go#L180)):*
> ```Go
> 
> 	// No results, was it due to unavailability or truncation?
> 	// TODO(irfansharif): Explore writing truncated state to raft engine.
> ```

I'm less sure about this one but it's probably a good idea to move TruncatedState too. 

---

*[pkg/storage/store.go, line 158 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-Ko3w1Pp2Piuv98TtmOL:-Ko3w1Pp2Piuv98TtmOM:b-hcre43) ([raw file](https://github.com/cockroachdb/cockroach/blob/1fc847963b49306c3ffe795ccd9df5a37505648b/pkg/storage/store.go#L158)):*
> ```Go
> )
> 
> // TODO(irfansharif): Changing this to a cluster setting instead makes it
> ```

Yes, this should be a cluster setting (which means that the global variables will have to be changed to functions). 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/16809#-:-Ko3wK7G_xvYfllkSycA:b-48shz0)*
<!-- Sent from Reviewable.io -->